Directory controller supports cross-table filter
Module configuration can specify filter which reference fields on other
entities. Filter specification in module config can have a new field for
configuration of this behavior called `join` which has the following
layout:

    join:
      ref:
        - joinField
        - joinAlias
      field: filterField

For example, entity `Approval` has a field `buyer` which is an
associated entity `Buyer` with a field `email`. You can fielter approval
based on buyer's email with this configuration:

    filters:
      buyerEmail:
        title: Email
        join:
          ref: [buyer, b]
          field: email